+++ src/CSC/model/transmission/co2_pipeline.jl
@@ -71,8 +71,8 @@ function co2_pipeline(EP::Model, inputs::Dict, setup::Dict)
 
 	println("CO2 Pipeline Module")
 
-    T = inputs["T"] # Model operating time steps
-    Z = inputs["Z"]  # Model demand zones - assumed to be same for CO2 and electricity
+    T = inputs["T"]::Int # Model operating time steps
+    Z = inputs["Z"]::Int  # Model demand zones - assumed to be same for CO2 and electricity
     INTERIOR_SUBPERIODS = inputs["INTERIOR_SUBPERIODS"]
     START_SUBPERIODS = inputs["START_SUBPERIODS"]
     hours_per_subperiod = inputs["hours_per_subperiod"]
