+++ src/GenX_extensions/write_energy_revenue.jl
@@ -5,8 +5,8 @@ Function for writing energy revenue from the different generation technologies.
 """
 function write_energy_revenue(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
 	dfGen = inputs["dfGen"]
-	G = inputs["G"]     # Number of resources (generators, storage, DR, and DERs)
-	T = inputs["T"]     # Number of time steps (hours)
+	G = inputs["G"]::Int     # Number of resources (generators, storage, DR, and DERs)
+	T = inputs["T"]::Int     # Number of time steps (hours)
 	FLEX = inputs["FLEX"]
 	NONFLEX = setdiff(collect(1:G), FLEX)
 	dfEnergyRevenue = DataFrame(Region = dfGen.region, Resource = inputs["RESOURCES"], Zone = dfGen.Zone, Cluster = dfGen.cluster, AnnualSum = Array{Float64}(undef, G),)
