+++ src/GenX_extensions/write_subsidy_revenue.jl
@@ -5,7 +5,7 @@ Function for reporting subsidy revenue earned if a generator specified `Min_Cap`
 """
 function write_subsidy_revenue(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
 	dfGen = inputs["dfGen"]
-	G = inputs["G"]
+	G = inputs["G"]::Int
 
 	dfSubRevenue = DataFrame(Region = dfGen[!, :region], Resource = inputs["RESOURCES"], Zone = dfGen[!, :Zone], Cluster = dfGen[!, :cluster], R_ID=dfGen[!, :R_ID], SubsidyRevenue = zeros(G))
 	MIN_CAP = dfGen[(dfGen[!, :Min_Cap_MW].>0), :R_ID]
