+++ src/HSC/model/g2p/h2_g2p.jl
@@ -33,8 +33,8 @@ function h2_g2p(EP::Model, inputs::Dict, setup::Dict)
     H2_G2P_COMMIT = inputs["H2_G2P_COMMIT"]
     H2_G2P_NO_COMMIT = inputs["H2_G2P_NO_COMMIT"]
     dfH2G2P = inputs["dfH2G2P"]  # Input H2 generation and storage data
-    Z = inputs["Z"]  # Model demand zones - assumed to be same for H2 and electricity
-    T = inputs["T"]     # Model operating time steps
+    Z = inputs["Z"]::Int  # Model demand zones - assumed to be same for H2 and electricity
+    T = inputs["T"]::Int     # Model operating time steps
 
     if !isempty(H2_G2P_COMMIT)
         EP = h2_g2p_commit(EP::Model, inputs::Dict, setup::Dict)
