+++ src/HSC/model/truck/h2_long_duration_truck.jl
@@ -45,7 +45,7 @@ function h2_long_duration_truck(EP::Model, inputs::Dict)
 
     print_and_log("H2 Long Duration Truck Module")
 
-    Z = inputs["Z"] # Number of zone locations
+    Z = inputs["Z"]::Int # Number of zone locations
     H2_TRUCK_TYPES = inputs["H2_TRUCK_TYPES"]
     dfPeriodMap = inputs["Period_Map"]
