+++ src/HSC/write_outputs/write_h2_balance_dual.jl
@@ -16,8 +16,8 @@ received this license file.  If not, see <http://www.gnu.org/licenses/>.
 
 function write_h2_balance_dual(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)
 	
-	T = inputs["T"]     # Number of time steps (hours)
-	Z = inputs["Z"]     # Number of zones
+	T = inputs["T"]::Int     # Number of time steps (hours)
+	Z = inputs["Z"]::Int     # Number of zones
 
 	# # Dual of storage level (state of charge) balance of each resource in each time step
 	dfH2BalanceDual = DataFrame(Zone = 1:Z)
