+++ src/HSC/write_outputs/write_h2_pipeline_level.jl
@@ -22,7 +22,7 @@ Function for reporting hydrogen storage level for each pipeline.
 function write_h2_pipeline_level(path::AbstractString, sep::AbstractString, inputs::Dict,setup::Dict, EP::Model)
 
     P = inputs["H2_P"]  # Number of H2 pipelines
-    T = inputs["T"]  # Model operating time steps
+    T = inputs["T"]::Int  # Model operating time steps
 
     p = zeros(P, T)
     dfH2PipelineLevel = DataFrame(Pipelines = 1:P)
