+++ src/HSC/write_outputs/write_h2_transmission_flow.jl
@@ -19,8 +19,8 @@ received this license file.  If not, see <http://www.gnu.org/licenses/>.
 """
 function write_h2_transmission_flow(path::AbstractString, sep::AbstractString, inputs::Dict,setup::Dict, EP::Model)
     
-    Z = inputs["Z"]     # Number of zones
-    T = inputs["T"]     # Number of time steps (hours)
+    Z = inputs["Z"]::Int     # Number of zones
+    T = inputs["T"]::Int     # Number of time steps (hours)
     
     dfH2TransmissionFlow = DataFrame(Time_Index=1:T)
     if setup["ModelH2Pipelines"] == 1
